cloud/network_server/requirements.txt
tornado
torndb
python-memcached
mysql-python
pymongo
json
jsonschema
tornadorpc
